+++ storage_drivers/ontap/ontap_san_economy.go
@@ -780,7 +780,9 @@ func (d *SANEconomyStorageDriver) DeleteBucketIfEmpty(bucketVol string) error {
 // Publish the volume to the host specified in publishInfo.  This method may or may not be running on the host
 // where the volume will be mounted, so it should limit itself to updating access rules, initiator groups, etc.
 // that require some host identity (but not locality) as well as storage controller API access.
-func (d *SANEconomyStorageDriver) Publish(name string, publishInfo *utils.VolumePublishInfo) error {
+func (d *SANEconomyStorageDriver) Publish(volConfig *storage.VolumeConfig, publishInfo *utils.VolumePublishInfo) error {
+
+	name := volConfig.InternalName
 
 	if d.Config.DebugTraceFlags["method"] {
 		fields := log.Fields{