+++ utils/utils_test.go
@@ -255,21 +255,21 @@ func TestGetYAMLTagWithSpaceCount(t *testing.T) {
 	assert.Equal(t, tag, "REPLACE-1")
 	assert.Equal(t, spaces, 0)
 
-	inputStringCopy := strings.ReplaceAll(inputString, tagWithSpaces, "#" + tagWithSpaces)
+	inputStringCopy := strings.ReplaceAll(inputString, tagWithSpaces, "#"+tagWithSpaces)
 
 	tagWithSpaces, tag, spaces = GetYAMLTagWithSpaceCount(inputStringCopy)
 	assert.Equal(t, tagWithSpaces, "  {REPLACE-2}\n")
 	assert.Equal(t, tag, "REPLACE-2")
 	assert.Equal(t, spaces, 2)
 
-	inputStringCopy = strings.ReplaceAll(inputStringCopy, tagWithSpaces, "#" + tagWithSpaces)
+	inputStringCopy = strings.ReplaceAll(inputStringCopy, tagWithSpaces, "#"+tagWithSpaces)
 
 	tagWithSpaces, tag, spaces = GetYAMLTagWithSpaceCount(inputStringCopy)
 	assert.Equal(t, tagWithSpaces, "    {REPLACE-3}\n")
 	assert.Equal(t, tag, "REPLACE-3")
 	assert.Equal(t, spaces, 4)
 
-	inputStringCopy = strings.ReplaceAll(inputStringCopy, tagWithSpaces, "#" + tagWithSpaces)
+	inputStringCopy = strings.ReplaceAll(inputStringCopy, tagWithSpaces, "#"+tagWithSpaces)
 
 	tagWithSpaces, tag, spaces = GetYAMLTagWithSpaceCount(inputStringCopy)
 	assert.Equal(t, tagWithSpaces, "    {REPLACE-4}\n")
@@ -294,11 +294,11 @@ func TestCountSpacesBeforeText(t *testing.T) {
 	log.Debug("Running TestCountSpacesBeforeText...")
 
 	type TextWithSpaces struct {
-		Text string
-		Spaces  int
+		Text   string
+		Spaces int
 	}
 
-	multipleTextsWithSpaces := []TextWithSpaces {
+	multipleTextsWithSpaces := []TextWithSpaces{
 		{"text", 0},
 		{"  text", 2},
 		{"  text  ", 2},
@@ -309,5 +309,49 @@ func TestCountSpacesBeforeText(t *testing.T) {
 	for _, textWithSpaces := range multipleTextsWithSpaces {
 		spaceCount := CountSpacesBeforeText(textWithSpaces.Text)
 		assert.Equal(t, spaceCount, textWithSpaces.Spaces)
+	}
+}
+
+func TestGetNFSVersionFromMountOptions(t *testing.T) {
+	log.Debug("Running TestGetNFSVersionFromMountOptions...")
+
+	defaultVersion := "3"
+	supportedVersions := []string{"3", "4", "4.1"}
+
+	var tests = []struct {
+		mountOptions      string
+		defaultVersion    string
+		supportedVersions []string
+		version           string
+		errNotNil         bool
+	}{
+		// Positive tests
+		{"", defaultVersion, supportedVersions, defaultVersion, false},
+		{"", defaultVersion, nil, defaultVersion, false},
+		{"vers=3", defaultVersion, supportedVersions, defaultVersion, false},
+		{"tcp, vers=3", defaultVersion, supportedVersions, defaultVersion, false},
+		{"-o hard,vers=4", defaultVersion, supportedVersions, "4", false},
+		{"nfsvers=3 , timeo=300", defaultVersion, supportedVersions, defaultVersion, false},
+		{"retry=1, nfsvers=4", defaultVersion, supportedVersions, "4", false},
+		{"retry=1", defaultVersion, supportedVersions, defaultVersion, false},
+		{"vers=4,minorversion=1", defaultVersion, supportedVersions, "4.1", false},
+		{"vers=4.0,minorversion=1", defaultVersion, supportedVersions, "4.1", false},
+		{"vers=2,vers=3", defaultVersion, supportedVersions, "3", false},
+		{"vers=5", defaultVersion, nil, "5", false},
+		{"minorversion=2,nfsvers=4.1", defaultVersion, supportedVersions, "4.1", false},
+
+		// Negative tests
+		{"vers=2", defaultVersion, supportedVersions, "2", true},
+		{"vers=3,minorversion=1", defaultVersion, supportedVersions, "3.1", true},
+		{"vers=4,minorversion=2", defaultVersion, supportedVersions, "4.2", true},
+		{"vers=4.1", defaultVersion, []string{"3", "4.2"}, "4.1", true},
+	}
+
+	for _, test := range tests {
+
+		version, err := GetNFSVersionFromMountOptions(test.mountOptions, test.defaultVersion, test.supportedVersions)
+
+		assert.Equal(t, test.version, version)
+		assert.Equal(t, test.errNotNil, err != nil)
 	}
 }